custom-elements/instantiating-custom-elements/custom-element-type-is-attribute.html

<!DOCTYPE html>
<html>
<head>
<title>Custom element type is the value of the is attribute of the custom element</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="The custom element type can be given to a custom element at the time of its instantation as the value of the is attribute of the custom element">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(tagName) {
        if (HTML5_DOCUMENT_ELEMENTS.indexOf(tagName) !== -1) {
            return;
        }
        var obj = doc.createElement(tagName);
        var name = 'x-a-' + tagName;
        doc.registerElement(name, {
            prototype: newCustomElementPrototype(obj.constructor.prototype, 'x-a'),
            extends: tagName
        });

        if (HTML5_TABLE_ELEMENTS.indexOf(tagName) !== -1) {
            doc.body.innerHTML =
                '<table>' +
                '<' + tagName + ' id="custom-element" is="' + name + '"></' + tagName + '>' +
                '</table>';
        } else {
            doc.body.innerHTML =
                '<' + tagName + ' id="custom-element" is="' + name + '"></' + tagName + '>';
        }
        var customElement = doc.querySelector('#custom-element');
        assert_equals(customElement.constructor.prototype.typeName, 'x-a',
            'Custom element type should be value of the IS attribute');
    });
}, 'Custom element type is the value of the IS attribute of the custom element');


test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(tagName) {
        if (HTML5_DOCUMENT_ELEMENTS.indexOf(tagName) !== -1) {
            return;
        }
        var obj = doc.createElement(tagName);
        var name = 'x-b-' + tagName;
        if (HTML5_TABLE_ELEMENTS.indexOf(tagName) !== -1) {
            doc.body.innerHTML =
                '<table>' +
                '<' + tagName + ' id="custom-element" is="' + name + '"></' + tagName + '>' +
                '</table>';
        } else {
            doc.body.innerHTML =
                '<' + tagName + ' id="custom-element" is="' + name + '"></' + tagName + '>';
        }
        doc.registerElement(name, {
            prototype: newCustomElementPrototype(obj.constructor.prototype, name),
            extends: tagName
        });
        var customElement = doc.querySelector('#custom-element');

        assert_equals(customElement.constructor.prototype.typeName, name,
            'Custom element type should be value of the IS attribute');
    });
}, 'Custom element type is the value of the IS attribute of the custom element. ' +
    'Custom element is unresolved at first');


test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(tagName) {
        if (HTML5_DOCUMENT_ELEMENTS.indexOf(tagName) !== -1) {
            return;
        }
        var obj = doc.createElement(tagName);
        var name = 'x-c-' + tagName;
        var GeneratedConstructor = doc.registerElement(name, {
            prototype: newHTMLElementPrototype(name),
            extends: tagName});
        var customElement = new GeneratedConstructor();

        assert_equals(customElement.getAttribute('is'), name,
            'Value of the IS attribute should be set to type');
    });
}, 'Value of the IS attribute should be set to type if local name is not the same as type');
</script>
</body>
</html>
